v0.96.25: 控件基类提炼 + 渲染器命名空间重构 + 编译警告清零
- 新增 TuiDisplayControl（CanFocus=false 收纳 15 纯展示控件）/ TuiBorderedControl（BorderStyle/BorderColor/GetBorderChars 统一 4 边框控件）；TuiPanel 删重复 switch 改委托 AnsiHelper（补全 Solid/Dotted 等样式）
- ToolRenderers → Renderers 命名空间统一；TuiMarkdown 移到 UI/TUI/；控件格式化统一
- 编译警告清零：CS2002 重复编译 / CS0675 符号扩展 / CS8601/8602/8604 null 修复
- 细节：toast 队列上限 5→10、列表 Home 到底、模型信息栏精简
- 自测 4720 全过

+++ WayCoder/Git/PackFile.cs
@@ -1,5 +1,4 @@
 using System.IO.Compression;
-using System.Runtime.InteropServices;
 using System.Security.Cryptography;
 using System.Text;
 
@@ -148,7 +147,7 @@ public static class PackFileReader
                 while ((c & 0x80) != 0)
                 {
                     c = pack[pos++];
-                    off = ((off + 1) << 7) | (c & 0x7F);
+                    off = ((off + 1) << 7) | (long)(c & 0x7F);
                 }
                 baseOffset = objOffset - off;
             }
@@ -295,7 +294,7 @@ public static class PackFileReader
             {
                 byte c = headBuf[pos++];
                 long off = c & 0x7F;
-                while ((c & 0x80) != 0) { c = headBuf[pos++]; off = ((off + 1) << 7) | (c & 0x7F); }
+                while ((c & 0x80) != 0) { c = headBuf[pos++]; off = ((off + 1) << 7) | (long)(c & 0x7F); }
                 baseOffset = objOffset - off;
             }
             else if (type == 7) // ref-delta：20 字节 base sha
@@ -609,7 +608,7 @@ public static class PackFileWriter
 
     static void WriteObjectHeader(Stream s, int type, long size)
     {
-        byte b = (byte)((type << 4) | (size & 0x0F));
+        byte b = (byte)(((long)type << 4) | (size & 0x0F));
         size >>= 4;
         if (size > 0) b |= 0x80;
         s.WriteByte(b);
